Sprawdzanie poprawności danych c.d.
Aby zapewnić poprawność wyświetlania maski uprawnień zmienono typ danych zmiennej rmask z Byte na Short (zmiany w bazie nie są konieczne).
TODO: rozwiązać problem milisekund w timestamp (nie działa wyszukiwanie typu =)

+++ AplikacjaDesktop_EE/src/java/Obiekty/MenuBar.java
@@ -32,7 +32,7 @@ public class MenuBar extends JMenuBar implements ActionListener {
 
     public void setRights(UserDTO user) {
 
-        Byte rmask = (user != null ? user.getRmask() : Obiekty.Def.LVL0);
+        Short rmask = (user != null ? user.getRmask() : Obiekty.Def.LVL0);
         JMenu menu, subMenu;
         JMenuItem menuItem;
